+++ dwds/test/build_daemon_evaluate_test.dart
@@ -126,6 +126,27 @@ void main() async {
             await setup.service.resume(isolate.id);
           });
 
+          test('with scope override is not supported yet', () async {
+            await onBreakPoint(isolate.id, mainScript, 'printLocal', () async {
+              var event = await stream.firstWhere(
+                  (event) => event.kind == EventKind.kPauseBreakpoint);
+
+              var object = await setup.service.evaluateInFrame(
+                  isolate.id, event.topFrame.index, 'MainClass(0)');
+
+              var param = object as InstanceRef;
+
+              expect(
+                  () => setup.service.evaluateInFrame(
+                        isolate.id,
+                        event.topFrame.index,
+                        't.toString()',
+                        scope: {'t': param.id},
+                      ),
+                  throwsRPCError);
+            });
+          });
+
           test('local', () async {
             await onBreakPoint(isolate.id, mainScript, 'printLocal', () async {
               var event = await stream.firstWhere(
@@ -476,6 +497,24 @@ void main() async {
           });
 
           tearDown(() async {});
+
+          test('with scope override', () async {
+            var library = isolate.rootLib;
+            var object = await setup.service
+                .evaluate(isolate.id, library.id, 'MainClass(0)');
+
+            var param = object as InstanceRef;
+            var result = await setup.service.evaluate(
+                isolate.id, library.id, 't.toString()',
+                scope: {'t': param.id});
+
+            expect(
+                result,
+                const TypeMatcher<InstanceRef>().having(
+                    (instance) => instance.valueAsString,
+                    'valueAsString',
+                    '0'));
+          });
 
           test('uses symbol from the same library', () async {
             var library = isolate.rootLib;